:construction: organizing model structure
