+++ client/src/models/teamModel.ts
@@ -1,0 +1,4 @@
+export interface ITeamSelectOption {
+  value: string;
+  label: string;
+}
